fixed remaining compiler warnings

+++ pgtklib/src/pgtk/Decoder.h
@@ -6,6 +6,6 @@ int pgtk_decoder_write(unsigned char* data, unsigned long long data_size, unsign
 
 unsigned long long pgtk_decode_v1(const Image_t* image, void* buffer, unsigned long long buffer_size, unsigned char threshold, unsigned char size);
 
-int pgtk_decode_marker(const Image_t* image, unsigned long long x, unsigned long long y, unsigned char threshold, unsigned char orientation, unsigned char* size, unsigned char* version);
+int pgtk_decode_marker(const Image_t* image, unsigned int x, unsigned int y, unsigned char threshold, unsigned char orientation, unsigned char* size, unsigned char* version);
 
 #endif
